NP-1715 Move ProjectsField component
