+++ src/pages/publication/description_tab/projects_field/ProjectOption.tsx
@@ -1,10 +1,10 @@
 import React, { FC } from 'react';
 import { AutocompleteRenderOptionState } from '@material-ui/lab';
-import { CristinProject } from '../../types/project.types';
 import { getProjectTitleParts } from './helpers';
-import { StyledFlexColumn } from '../styled/Wrappers';
 import ProjectInstitutions from './ProjectInstitutions';
 import ProjectTitle from './ProjectTitle';
+import { CristinProject } from '../../../../types/project.types';
+import { StyledFlexColumn } from '../../../../components/styled/Wrappers';
 
 interface ProjectOptionProps {
   project: CristinProject;
