+++ src/pages/publication/description_tab/projects_field/index.ts
@@ -1,0 +1,1 @@
+export { default as ProjectsField } from './ProjectsField';
